refact : split funcs. expand.c -> expand_util_1.c
