+++ jaemjeon_jisookim_minishell/main/Makefile
@@ -36,7 +36,7 @@ SRC_UTILS		:= $(addprefix $(DIR_UTILS)/, $(SRC_UTILS))
 SRC_MAIN		:= main.c init.c
 SRC_MAIN		:= $(addprefix $(DIR_MAIN)/, $(SRC_MAIN))
 
-SRC_EXPAND		:= expand.c
+SRC_EXPAND		:= expand.c expand_util_1.c
 SRC_EXPAND		:= $(addprefix $(DIR_EXPAND)/, $(SRC_EXPAND))
 
 SRC_PARSE		:= parse.c cmd_adt_1.c parse_util_1.c
